+++ 2-advanced/dubbo-samples-context/dubbo-samples-context-consumer/src/main/resources/log4j.properties
